% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{assertOptOutFile}
\alias{assertOptOutFile}
\title{Assert option is valid output file name}
\usage{
assertOptOutFile(optName, file, force = NA)
}
\arguments{
\item{optName}{The name of the option (used in error messages).}

\item{file}{The file name (path) to check for useability.}

\item{force}{If set TRUE, will delete \code{file} if it exists (with a warning)
and create any needed directories (silently). If FALSE or NA, these will be
errors, with only the text of the error differing. NA, the default, means
no 'force' option is mentioned, while FALSE implies \code{--force} is available
as an option, but was not set.}
}
\value{
Exits with error or returns TRUE
}
\description{
Checks that a value provided as an option is suitable for use as an output
file path. Can be relative or absolute. It is an error if the file exists
or its directory does not,  unless "force" is set TRUE, in which case the
directory will be (silently) created or the pre-existing file will be
deleted  (with a warning). Failure to create a needed directory or delete
a pre-existing file also results in an error.
}
