% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.R
\name{groupedIntersect}
\alias{groupedIntersect}
\title{Intersect groups of ranges with common target range set}
\usage{
groupedIntersect(x, y, group, simplify = TRUE)
}
\arguments{
\item{x}{The grouped (query) ranges}

\item{y}{The common target (subject) ranges}

\item{group}{The groupings, must be a vector of values the same length
as x (query ranges). Can be specified as a vector, as a list with one
element, or as the name of a meta-data column in x (query ranges). See
\code{simplify} for additional info.}

\item{simplify}{Set FALSE to return a GRangesList, by group value. By default
simplify is TRUE and a concatenated GRanges object will be returned, with
a metadata column giving the group value named based on the format of 'group'
\itemize{
\item vector: column will be named "group".
\item list: column will be named names(group)\link{1}.
\item column name: column will have the same name.
}}
}
\value{
A GRanges object or GRangesList object giving the intersection of
all 'y' (subject) ranges with each group of 'x' (query ranges) as specified
by "group". If either is NA, returns NA. If either is empty, returns an
empty GRangesList object, or an empty GRanges object if \code{simplify} is \code{TRUE}.
}
\description{
Intersect groups of ranges with common target range set
}
