% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{loadBed}
\alias{loadBed}
\title{Load a bed-like file as a GRanges object}
\usage{
loadBed(file, colnames = c("seqname", "start", "end"), oneBased = FALSE, ...)
}
\arguments{
\item{file}{The bed file to load}

\item{colnames}{Character vector that specifies how to map loaded data to a
gr object. Only the specified columns (in order) will be loaded; any
following unspecified columns will be dropped, as will any columns whose
names are NA. Three column are required "seqnames", "start", and
"end". The default is for these to be the first three columns. If "strand"
is named, this column will be used as the strand, otherwise the strand for
all loci will be "*". If other columns are named, those will be used as
metadata columns, w}

\item{oneBased}{By default, input start values are assumed to be zero-based,
as per the bed specification. Set \code{TRUE} if input start values are one-based.}

\item{...}{Extra parameters for read.table when reading the file, e.g.
setting \code{skip= 1} or \code{header= TRUE} to skip loading headings (file headings
are never used; \code{colnames} must always be specified.)}
}
\value{
GenomicRanges::GRanges object
}
\description{
Loads a bed-like file of space or tab-delimited genomic location data into a
genomic ranges object. Columns are identified based on provided column names.
Start positions are assumed to be 0 based in the file. Genomic ranges objects
use 1-based positions.
}
\examples{
# Sample bed detail file
bed6File <- system.file( "extdata", "bed6.bed", package = "LowCover" )
readLines(bed6File)

# Load location information only, use a (default) strand of '*'
gr <- loadBed( bed6File )
gr

# Load all data, including strand and two meta-data columns, 'score' and 'id'
mapCols <- c("seqname", "start", "end", "strand", "score", "id")
gr <- loadBed( bed6File, colnames= mapCols)
gr

# Load required columns + score only. Use (default) strand of '*'
mapCols <- c("seqname", "start", "end", NA, "score")
gr <- loadBed( bed6File, colnames= mapCols)
gr

# Load bed-like file with header and 1 based location column info
bed6LikeFile <- system.file( "extdata", "bed6like.txt", package = "LowCover" )
readLines(bed6LikeFile)

mapCols <- c("seqname", "start", "end", "strand", "score", "id")
gr <- loadBed( bed6File, colnames= mapCols, oneBased= TRUE, skip= 1)
gr

}
