% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LowCover.R
\name{lowcoverApp}
\alias{lowcoverApp}
\title{Run lowcover as an application}
\usage{
lowcoverApp(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{Character vector of command line parameters, by default will
get them from the command line used to start the R session in which this
runs in the usual way.}
}
\value{
N/A - run as an app that saves everything to files
}
\description{
Run lowcover as an application
}
