% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{parseCLI}
\alias{parseCLI}
\title{Parse the command line options and arguments.}
\usage{
parseCLI(args = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{args}{The vector or command line tokens to parse, defaults to
retrieving them itself.}
}
\value{
A list of the parsed command line options and arguments. Contains an
unnamed option as the first element due to undocumented behavior in
argparser.
}
\description{
Returns an unvalidated list of options by long name. (-h,--help) and
(-V,--version) are parsed manually, so only help=TRUE will be present if help
specified, or if no help asked for but versions is specified, only
version=TRUE will be present.
}
