% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveBed}
\alias{saveBed}
\title{Save a Granges object as a bed file.}
\usage{
saveBed(x, file, strand = TRUE, oneBased = FALSE, col.names = FALSE, ...)
}
\arguments{
\item{x}{GenomicRanges::Granges object to write to a bed file}

\item{file}{Filename or connection to write to.}

\item{strand}{By default, writes a strand column. Set this TRUE to skip.}

\item{oneBased}{By default, writes 0-based start locations. Set this TRUE to
use one-based starts}

\item{col.names}{By default, writes no header. Pass a vector of names or set
this TRUE to use the default "as.data.frame" GRanges column names.}

\item{...}{Additional parameters to write.table}
}
\value{
The result of calling write.table(), which is undocumented.
}
\description{
Saves a Granges object to a tab-delimited file with "seqnames", "start",
"end", and "strand" as the first four columns, followed by all metadata
columns. Will convert to 0 based starts and will not include column names
unless parameters indicate otherwise. Strand can be excluded if \code{strand} is
set false. To save compressed output, use a connection like gzcon instead of
a file. You will have to close it yourself after use.
}
\examples{
gr <- GenomicRanges::GRanges(
  seqnames= c( "chr1", "chr1", "chr1", "chr2", "chr10" ),
  ranges <- IRanges::IRanges(
    start=    c( 1L, 11L, 11L, 1L, 16L ),
    end=      c( 10L, 20L, 20L, 249000000L, 15L )),
    strand=   c( "*", "+", "+", "*", "-" ),
    score=    c( 1.1, 2.2, 3.3, 4.4, 5.5 ),
    id=       c( "bob", "sue", "tom", "april", "alice" ))

bedFile <- tempfile( fileext= ".bed" )
saveBed( gr, bedFile )
readLines( bedFile )

bed3File <- tempfile( fileext= ".bed" )
grMin <- gr
GenomicRanges::mcols( gr ) <- NULL
saveBed( grMin, bed3File, strand= FALSE )
readLines( bed3File )

tsvFile <- tempfile( fileext= ".tsv" )
saveBed( gr, tsvFile, oneBased=TRUE, col.names = TRUE )

}
