% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{validateOptions}
\alias{validateOptions}
\title{Validate CLI options}
\usage{
validateOptions(rawOpts, unvalidated = c("warning", "error", "ok", "ignore"))
}
\arguments{
\item{rawOpts}{The unvalidated options list}

\item{unvalidated}{What to do with unvalidated options:
\itemize{
\item \code{warning} - By default if an option is not validated, it is used as is, but
a warning will be generated.
\item \code{error}  - Unvalidated options will trigger an error.
\item \code{ok} - Unvalidated options will silently be used as is.
\item \code{ignore} - Unvalidated options will silently be ignored.
}}
}
\value{
The list of validated options, including any unvalidated options if
\code{unvalidated} is "warning" or "ok".
}
\description{
Validates (and possibly transforms) command line options. Halts with error on
the first invalid option found, or returns a list of valid options and their
post-validation values. The list of raw options, including any invalid opts,
is included in the returned list as \code{rawOpts} and returns a list of valid
options.How unknown options are handled depends on \code{unvalidated}.
}
