% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowcover-package.R
\docType{package}
\name{lowcover-package}
\alias{lowcover}
\alias{lowcover-package}
\title{lowcover: Low coverage reporting}
\description{
Reports low coverage genomic regions based on mosdepth.
}
\author{
\strong{Maintainer}: Stuart R. Jefferys \email{srjefferys@gmail.com} (\href{https://orcid.org/0000-0001-6259-085X}{ORCID})

}
\keyword{internal}
